+++ src/main/java/com/exemple/quiz_app/auth/repository/UserRepository.java
@@ -5,13 +5,13 @@ import com.exemple.quiz_app.auth.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
+
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, BigInteger> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
     boolean existsByEmail(String email);
-    boolean existsById(BigInteger id);
+    boolean existsById(Long id);
 }
